backend/requirements.txt
flask==2.3.3
flask-restful==0.3.10
flask-cors==4.0.0
pymongo==4.5.0
python-dotenv==1.0.0
pydub==0.25.1
SpeechRecognition==3.10.0
gTTS==2.3.2
dnspython==2.4.2
requests==2.31.0
beautifulsoup4==4.12.2 